+++ test/testdataimage.jl
@@ -1,4 +1,5 @@
 @testset "DataImage" begin
+   
     datamat = [
         1.0 2 3
         4 5 6